move command_queue to impl
