+++ tests/tt_metal/tt_metal/common/matmul_test_utils.hpp
@@ -12,7 +12,7 @@
 #include <tt-metalium/tilize_utils.hpp>
 #include <tt-metalium/distributed.hpp>
 #include "hostdevcommon/common_values.hpp"
-#include <tt-metalium/command_queue.hpp>
+#include "tt_metal/impl/dispatch/command_queue.hpp"
 #include "llrt.hpp"
 
 namespace tt::tt_metal {
