+++ tests/tt_metal/tt_metal/common/mesh_dispatch_fixture.hpp
@@ -12,7 +12,7 @@
 #include "hostdevcommon/common_values.hpp"
 #include "tt_metal/test_utils/env_vars.hpp"
 #include <tt-metalium/program.hpp>
-#include <tt-metalium/command_queue.hpp>
+#include "tt_metal/impl/dispatch/command_queue.hpp"
 #include <tt-metalium/device.hpp>
 #include <tt-metalium/device_pool.hpp>
 #include <tt-metalium/distributed.hpp>
