+++ tests/tt_metal/tt_metal/dispatch/dispatch_event/test_events.cpp
@@ -6,7 +6,7 @@
 #include <fmt/base.h>
 #include <stddef.h>
 #include <stdint.h>
-#include <tt-metalium/command_queue.hpp>
+#include "tt_metal/impl/dispatch/command_queue.hpp"
 #include <tt-metalium/event.hpp>
 #include <tt-metalium/host_api.hpp>
 #include <future>
