+++ tt_metal/api/tt-metalium/mesh_command_queue.hpp
@@ -19,7 +19,6 @@
 
 #include <tt_stl/span.hpp>
 #include <tt-metalium/buffer.hpp>
-#include <tt-metalium/command_queue.hpp>
 #include <tt-metalium/core_coord.hpp>
 #include <tt-metalium/mesh_buffer.hpp>
 #include <tt-metalium/mesh_coord.hpp>
