+++ tt_metal/distributed/fd_mesh_command_queue.hpp
@@ -6,7 +6,7 @@
 
 #include "mesh_command_queue_base.hpp"
 
-#include <tt-metalium/command_queue.hpp>
+#include "tt_metal/impl/dispatch/command_queue.hpp"
 
 #include "tt_metal/common/multi_producer_single_consumer_queue.hpp"
 #include "dispatch/cq_shared_state.hpp"
