+++ tt_metal/impl/device/device.cpp
@@ -32,7 +32,7 @@
 #include "allocator.hpp"
 #include "allocator_types.hpp"
 #include <tt_stl/assert.hpp>
-#include "command_queue.hpp"
+#include "dispatch/command_queue.hpp"
 #include "dispatch/command_queue_common.hpp"
 #include "common/core_assignment.hpp"
 #include "program/program_impl.hpp"
